+++ docs/changelog.rst
@@ -10,8 +10,9 @@ Changelog
 - Added Python 3.4 support.
 - Added :meth:`Connection.wait_closed`.
 - Added :meth:`Connection.is_closing`.
-- Added :attr:`Connection.remote_host`.
+- Added :attr:`Connection.remote_hostname`.
 - Added ``auto_link`` parameter to :func:`connect`, :func:`start_server` and :func:`run_app`.
+- Added ``remote_hostname`` parameter to :func:`connect`.
 - Connection open and close log messages demoted from ``INFO`` to ``DEBUG`` level.